requirements.txt
django==3.1.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3